move open_file to model_methods
